test(timeline): recategorize and clean up browser tests
